+++ decor/applications/lobsters/microbenchmarks/queriers/user.rs
@@ -5,14 +5,14 @@ use mysql::prelude::*;
 use std::*;
 //use log::{warn, debug};
 
-pub fn get_profile(db: &mut mysql::Conn, uid: u32) -> Result<(), mysql::Error> {
-    let uid : u32 = db.query(format!(
+pub fn get_profile(db: &mut mysql::Conn, uid: u64) -> Result<(), mysql::Error> {
+    let uid : u64 = db.query(format!(
             "SELECT `users`.id FROM `users` \
              WHERE `users`.`username` = {}",
             (format!("\'user{}\'", uid))
         ))?[0];
 
-    let rows : Vec<(u32, u32)> = db.query(format!(
+    let rows : Vec<(u64, u64)> = db.query(format!(
             "SELECT  `tags`.`id`, COUNT(*) AS `count` FROM `taggings` \
              INNER JOIN `tags` ON `taggings`.`tag_id` = `tags`.`id` \
              INNER JOIN `stories` ON `stories`.`id` = `taggings`.`story_id` \
@@ -24,7 +24,7 @@ pub fn get_profile(db: &mut mysql::Conn, uid: u32) -> Result<(), mysql::Error> {
         )?;
 
     if !rows.is_empty() {
-        let tag : u32 = rows[0].0;
+        let tag : u64 = rows[0].0;
         db.query_drop(format!(
             "SELECT  `tags`.* \
              FROM `tags` \
@@ -53,12 +53,12 @@ pub fn get_profile(db: &mut mysql::Conn, uid: u32) -> Result<(), mysql::Error> {
 } 
 
 
-pub fn resubscribe(db: &mut mysql::Conn, uid: u32) -> Result<(), mysql::Error> {
+pub fn resubscribe(db: &mut mysql::Conn, uid: u64) -> Result<(), mysql::Error> {
     db.query_drop(format!("RESUBSCRIBE UID {}", uid))?;
     Ok(())
 }
 
-pub fn unsubscribe(db: &mut mysql::Conn, uid: u32) -> Result<(), mysql::Error> {
+pub fn unsubscribe(db: &mut mysql::Conn, uid: u64) -> Result<(), mysql::Error> {
     db.query_drop(format!("UNSUBSCRIBE UID {}", uid))?;
     Ok(())
 }
